Location, customer and sales representative mapper's unit tests. Change visit percentage for double value in different customers dtos

+++ src/main/java/com/test/banshee/dto/customer/GetCustomerDTO.java
@@ -32,7 +32,7 @@ public class GetCustomerDTO {
 
     private BigDecimal availableCredit;
 
-    private BigDecimal visitPercentage;
+    private Double visitPercentage;
 
     private List<GetVisitDTO> visits;
